refactor registry接入方式
